컷신 페이드 효과 추가
