+++ Assets/Prefabs/MainWorld/Curtain.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fa31987a2e74fca48ab1b97c8582726b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
