+++ Assets/Scripts/MainWorld/FadeCurtain.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c1ff2f156e0ce1c449599e46e8b4fa11
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
